Splitted tests.py into separate files and folders
